agent-server/pyproject.toml
[project]
name = "agent-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]
